refactoring persistent render pass and framebuffer
