+++ cocos/rendering/custom/render-graph.ts
@@ -31,7 +31,7 @@
 import { AdjI, AdjacencyGraph, BidirectionalGraph, ComponentGraph, ED, InEI, MutableGraph, MutableReferenceGraph, NamedGraph, OutE, OutEI, PolymorphicGraph, PropertyGraph, PropertyMap, ReferenceGraph, UuidGraph, VertexListGraph, directional, parallel, reindexEdgeList, traversal } from './graph';
 import { Material } from '../../asset/assets';
 import { Camera } from '../../render-scene/scene/camera';
-import { AccessFlagBit, Buffer, ClearFlagBit, Color, Format, Framebuffer, SampleCount, Sampler, SamplerInfo, Swapchain, Texture, TextureFlagBit, Viewport } from '../../gfx';
+import { AccessFlagBit, Buffer, ClearFlagBit, Color, Format, Framebuffer, RenderPass, SampleCount, Sampler, SamplerInfo, Swapchain, Texture, TextureFlagBit, Viewport } from '../../gfx';
 import { ComputeView, CopyPair, LightInfo, MovePair, QueueHint, RasterView, ResourceDimension, ResourceFlags, ResourceResidency, SceneFlags } from './types';
 
 export class ResourceDesc {
@@ -85,6 +85,341 @@ export class ManagedTexture {
 
 export class ManagedResource {
     unused = 0;
+}
+
+export class RasterSubpass {
+    readonly rasterViews: Map<string, RasterView> = new Map<string, RasterView>();
+    readonly computeViews: Map<string, ComputeView[]> = new Map<string, ComputeView[]>();
+}
+
+//=================================================================
+// SubpassGraph
+//=================================================================
+// Graph Concept
+export class SubpassGraphVertex {
+    constructor () {
+    }
+    readonly _outEdges: OutE[] = [];
+    readonly _inEdges: OutE[] = [];
+}
+
+//-----------------------------------------------------------------
+// PropertyGraph Concept
+export class SubpassGraphNameMap implements PropertyMap {
+    constructor (readonly names: string[]) {
+        this._names = names;
+    }
+    get (v: number): string {
+        return this._names[v];
+    }
+    set (v: number, names: string): void {
+        this._names[v] = names;
+    }
+    readonly _names: string[];
+}
+
+export class SubpassGraphSubpassMap implements PropertyMap {
+    constructor (readonly subpasses: RasterSubpass[]) {
+        this._subpasses = subpasses;
+    }
+    get (v: number): RasterSubpass {
+        return this._subpasses[v];
+    }
+    readonly _subpasses: RasterSubpass[];
+}
+
+//-----------------------------------------------------------------
+// ComponentGraph Concept
+export const enum SubpassGraphComponent {
+    Name,
+    Subpass,
+}
+
+export interface SubpassGraphComponentType {
+    [SubpassGraphComponent.Name]: string;
+    [SubpassGraphComponent.Subpass]: RasterSubpass;
+}
+
+export interface SubpassGraphComponentPropertyMap {
+    [SubpassGraphComponent.Name]: SubpassGraphNameMap;
+    [SubpassGraphComponent.Subpass]: SubpassGraphSubpassMap;
+}
+
+//-----------------------------------------------------------------
+// SubpassGraph Implementation
+export class SubpassGraph implements BidirectionalGraph
+, AdjacencyGraph
+, VertexListGraph
+, MutableGraph
+, PropertyGraph
+, NamedGraph
+, ComponentGraph {
+    //-----------------------------------------------------------------
+    // Graph
+    // type vertex_descriptor = number;
+    nullVertex (): number { return 0xFFFFFFFF; }
+    // type edge_descriptor = ED;
+    readonly directed_category: directional = directional.bidirectional;
+    readonly edge_parallel_category: parallel = parallel.allow;
+    readonly traversal_category: traversal = traversal.incidence
+        | traversal.bidirectional
+        | traversal.adjacency
+        | traversal.vertex_list;
+    //-----------------------------------------------------------------
+    // IncidenceGraph
+    // type out_edge_iterator = OutEI;
+    // type degree_size_type = number;
+    edge (u: number, v: number): boolean {
+        for (const oe of this._vertices[u]._outEdges) {
+            if (v === oe.target as number) {
+                return true;
+            }
+        }
+        return false;
+    }
+    source (e: ED): number {
+        return e.source as number;
+    }
+    target (e: ED): number {
+        return e.target as number;
+    }
+    outEdges (v: number): OutEI {
+        return new OutEI(this._vertices[v]._outEdges.values(), v);
+    }
+    outDegree (v: number): number {
+        return this._vertices[v]._outEdges.length;
+    }
+    //-----------------------------------------------------------------
+    // BidirectionalGraph
+    // type in_edge_iterator = InEI;
+    inEdges (v: number): InEI {
+        return new InEI(this._vertices[v]._inEdges.values(), v);
+    }
+    inDegree (v: number): number {
+        return this._vertices[v]._inEdges.length;
+    }
+    degree (v: number): number {
+        return this.outDegree(v) + this.inDegree(v);
+    }
+    //-----------------------------------------------------------------
+    // AdjacencyGraph
+    // type adjacency_iterator = AdjI;
+    adjacentVertices (v: number): AdjI {
+        return new AdjI(this, this.outEdges(v));
+    }
+    //-----------------------------------------------------------------
+    // VertexListGraph
+    vertices (): IterableIterator<number> {
+        return this._vertices.keys();
+    }
+    numVertices (): number {
+        return this._vertices.length;
+    }
+    //-----------------------------------------------------------------
+    // EdgeListGraph
+    numEdges (): number {
+        let numEdges = 0;
+        for (const v of this.vertices()) {
+            numEdges += this.outDegree(v);
+        }
+        return numEdges;
+    }
+    //-----------------------------------------------------------------
+    // MutableGraph
+    clear (): void {
+        // ComponentGraph
+        this._names.length = 0;
+        this._subpasses.length = 0;
+        // Graph Vertices
+        this._vertices.length = 0;
+    }
+    addVertex (
+        name: string,
+        subpass: RasterSubpass,
+    ): number {
+        const vert = new SubpassGraphVertex();
+        const v = this._vertices.length;
+        this._vertices.push(vert);
+        this._names.push(name);
+        this._subpasses.push(subpass);
+        return v;
+    }
+    clearVertex (v: number): void {
+        const vert = this._vertices[v];
+        // clear out edges
+        for (const oe of vert._outEdges) {
+            const target = this._vertices[oe.target as number];
+            for (let i = 0; i !== target._inEdges.length;) { // remove all edges
+                if (target._inEdges[i].target === v) {
+                    target._inEdges.splice(i, 1);
+                } else {
+                    ++i;
+                }
+            }
+        }
+        vert._outEdges.length = 0;
+
+        // clear in edges
+        for (const ie of vert._inEdges) {
+            const source = this._vertices[ie.target as number];
+            for (let i = 0; i !== source._outEdges.length;) { // remove all edges
+                if (source._outEdges[i].target === v) {
+                    source._outEdges.splice(i, 1);
+                } else {
+                    ++i;
+                }
+            }
+        }
+        vert._inEdges.length = 0;
+    }
+    removeVertex (u: number): void {
+        this._vertices.splice(u, 1);
+        this._names.splice(u, 1);
+        this._subpasses.splice(u, 1);
+
+        const sz = this._vertices.length;
+        if (u === sz) {
+            return;
+        }
+
+        for (let v = 0; v !== sz; ++v) {
+            const vert = this._vertices[v];
+            reindexEdgeList(vert._outEdges, u);
+            reindexEdgeList(vert._inEdges, u);
+        }
+    }
+    addEdge (u: number, v: number): ED | null {
+        // update in/out edge list
+        this._vertices[u]._outEdges.push(new OutE(v));
+        this._vertices[v]._inEdges.push(new OutE(u));
+        return new ED(u, v);
+    }
+    removeEdges (u: number, v: number): void {
+        const source = this._vertices[u];
+        // remove out edges of u
+        for (let i = 0; i !== source._outEdges.length;) { // remove all edges
+            if (source._outEdges[i].target === v) {
+                source._outEdges.splice(i, 1);
+            } else {
+                ++i;
+            }
+        }
+        // remove in edges of v
+        const target = this._vertices[v];
+        for (let i = 0; i !== target._inEdges.length;) { // remove all edges
+            if (target._inEdges[i].target === u) {
+                target._inEdges.splice(i, 1);
+            } else {
+                ++i;
+            }
+        }
+    }
+    removeEdge (e: ED): void {
+        const u = e.source as number;
+        const v = e.target as number;
+        const source = this._vertices[u];
+        for (let i = 0; i !== source._outEdges.length;) {
+            if (source._outEdges[i].target === v) {
+                source._outEdges.splice(i, 1);
+                break; // remove one edge
+            } else {
+                ++i;
+            }
+        }
+        const target = this._vertices[v];
+        for (let i = 0; i !== target._inEdges.length;) {
+            if (target._inEdges[i].target === u) {
+                target._inEdges.splice(i, 1);
+                break; // remove one edge
+            } else {
+                ++i;
+            }
+        }
+    }
+    //-----------------------------------------------------------------
+    // NamedGraph
+    vertexName (v: number): string {
+        return this._names[v];
+    }
+    vertexNameMap (): SubpassGraphNameMap {
+        return new SubpassGraphNameMap(this._names);
+    }
+    //-----------------------------------------------------------------
+    // PropertyGraph
+    get (tag: string): SubpassGraphNameMap | SubpassGraphSubpassMap {
+        switch (tag) {
+        // Components
+        case 'Name':
+            return new SubpassGraphNameMap(this._names);
+        case 'Subpass':
+            return new SubpassGraphSubpassMap(this._subpasses);
+        default:
+            throw Error('property map not found');
+        }
+    }
+    //-----------------------------------------------------------------
+    // ComponentGraph
+    component<T extends SubpassGraphComponent> (id: T, v: number): SubpassGraphComponentType[T] {
+        switch (id) {
+        case SubpassGraphComponent.Name:
+            return this._names[v] as SubpassGraphComponentType[T];
+        case SubpassGraphComponent.Subpass:
+            return this._subpasses[v] as SubpassGraphComponentType[T];
+        default:
+            throw Error('component not found');
+        }
+    }
+    componentMap<T extends SubpassGraphComponent> (id: T): SubpassGraphComponentPropertyMap[T] {
+        switch (id) {
+        case SubpassGraphComponent.Name:
+            return new SubpassGraphNameMap(this._names) as SubpassGraphComponentPropertyMap[T];
+        case SubpassGraphComponent.Subpass:
+            return new SubpassGraphSubpassMap(this._subpasses) as SubpassGraphComponentPropertyMap[T];
+        default:
+            throw Error('component map not found');
+        }
+    }
+    getName (v: number): string {
+        return this._names[v];
+    }
+    setName (v: number, value: string) {
+        this._names[v] = value;
+    }
+    getSubpass (v: number): RasterSubpass {
+        return this._subpasses[v];
+    }
+
+    readonly components: string[] = ['Name', 'Subpass'];
+    readonly _vertices: SubpassGraphVertex[] = [];
+    readonly _names: string[] = [];
+    readonly _subpasses: RasterSubpass[] = [];
+}
+
+export class RasterPass {
+    readonly rasterViews: Map<string, RasterView> = new Map<string, RasterView>();
+    readonly computeViews: Map<string, ComputeView[]> = new Map<string, ComputeView[]>();
+    readonly subpassGraph: SubpassGraph = new SubpassGraph();
+    width = 0;
+    height = 0;
+    readonly viewport: Viewport = new Viewport();
+    versionName = '';
+    version = 0;
+    showStatistics = false;
+}
+
+export class PersistentRenderPassAndFramebuffer {
+    constructor (renderPass: RenderPass, framebuffer: Framebuffer) {
+        this.renderPass = renderPass;
+        this.framebuffer = framebuffer;
+    }
+    /*refcount*/ renderPass: RenderPass;
+    /*refcount*/ framebuffer: Framebuffer;
+    readonly clearColors: Color[] = [];
+    clearDepth = 0;
+    clearStencil = 0;
+    refCount = 1;
+    hash = 0;
+    version = 0;
 }
 
 //=================================================================
@@ -324,6 +659,7 @@ export class ResourceGraph implements BidirectionalGraph
     // MutableGraph
     clear (): void {
         // Members
+        this.renderPasses.clear();
         this.nextFenceValue = 0;
         this.version = 0;
         // UuidGraph
@@ -705,328 +1041,9 @@ export class ResourceGraph implements BidirectionalGraph
     readonly _states: ResourceStates[] = [];
     readonly _samplerInfo: SamplerInfo[] = [];
     readonly _valueIndex: Map<string, number> = new Map<string, number>();
+    readonly renderPasses: Map<string, PersistentRenderPassAndFramebuffer> = new Map<string, PersistentRenderPassAndFramebuffer>();
     nextFenceValue = 0;
     version = 0;
-}
-
-export class RasterSubpass {
-    readonly rasterViews: Map<string, RasterView> = new Map<string, RasterView>();
-    readonly computeViews: Map<string, ComputeView[]> = new Map<string, ComputeView[]>();
-}
-
-//=================================================================
-// SubpassGraph
-//=================================================================
-// Graph Concept
-export class SubpassGraphVertex {
-    constructor () {
-    }
-    readonly _outEdges: OutE[] = [];
-    readonly _inEdges: OutE[] = [];
-}
-
-//-----------------------------------------------------------------
-// PropertyGraph Concept
-export class SubpassGraphNameMap implements PropertyMap {
-    constructor (readonly names: string[]) {
-        this._names = names;
-    }
-    get (v: number): string {
-        return this._names[v];
-    }
-    set (v: number, names: string): void {
-        this._names[v] = names;
-    }
-    readonly _names: string[];
-}
-
-export class SubpassGraphSubpassMap implements PropertyMap {
-    constructor (readonly subpasses: RasterSubpass[]) {
-        this._subpasses = subpasses;
-    }
-    get (v: number): RasterSubpass {
-        return this._subpasses[v];
-    }
-    readonly _subpasses: RasterSubpass[];
-}
-
-//-----------------------------------------------------------------
-// ComponentGraph Concept
-export const enum SubpassGraphComponent {
-    Name,
-    Subpass,
-}
-
-export interface SubpassGraphComponentType {
-    [SubpassGraphComponent.Name]: string;
-    [SubpassGraphComponent.Subpass]: RasterSubpass;
-}
-
-export interface SubpassGraphComponentPropertyMap {
-    [SubpassGraphComponent.Name]: SubpassGraphNameMap;
-    [SubpassGraphComponent.Subpass]: SubpassGraphSubpassMap;
-}
-
-//-----------------------------------------------------------------
-// SubpassGraph Implementation
-export class SubpassGraph implements BidirectionalGraph
-, AdjacencyGraph
-, VertexListGraph
-, MutableGraph
-, PropertyGraph
-, NamedGraph
-, ComponentGraph {
-    //-----------------------------------------------------------------
-    // Graph
-    // type vertex_descriptor = number;
-    nullVertex (): number { return 0xFFFFFFFF; }
-    // type edge_descriptor = ED;
-    readonly directed_category: directional = directional.bidirectional;
-    readonly edge_parallel_category: parallel = parallel.allow;
-    readonly traversal_category: traversal = traversal.incidence
-        | traversal.bidirectional
-        | traversal.adjacency
-        | traversal.vertex_list;
-    //-----------------------------------------------------------------
-    // IncidenceGraph
-    // type out_edge_iterator = OutEI;
-    // type degree_size_type = number;
-    edge (u: number, v: number): boolean {
-        for (const oe of this._vertices[u]._outEdges) {
-            if (v === oe.target as number) {
-                return true;
-            }
-        }
-        return false;
-    }
-    source (e: ED): number {
-        return e.source as number;
-    }
-    target (e: ED): number {
-        return e.target as number;
-    }
-    outEdges (v: number): OutEI {
-        return new OutEI(this._vertices[v]._outEdges.values(), v);
-    }
-    outDegree (v: number): number {
-        return this._vertices[v]._outEdges.length;
-    }
-    //-----------------------------------------------------------------
-    // BidirectionalGraph
-    // type in_edge_iterator = InEI;
-    inEdges (v: number): InEI {
-        return new InEI(this._vertices[v]._inEdges.values(), v);
-    }
-    inDegree (v: number): number {
-        return this._vertices[v]._inEdges.length;
-    }
-    degree (v: number): number {
-        return this.outDegree(v) + this.inDegree(v);
-    }
-    //-----------------------------------------------------------------
-    // AdjacencyGraph
-    // type adjacency_iterator = AdjI;
-    adjacentVertices (v: number): AdjI {
-        return new AdjI(this, this.outEdges(v));
-    }
-    //-----------------------------------------------------------------
-    // VertexListGraph
-    vertices (): IterableIterator<number> {
-        return this._vertices.keys();
-    }
-    numVertices (): number {
-        return this._vertices.length;
-    }
-    //-----------------------------------------------------------------
-    // EdgeListGraph
-    numEdges (): number {
-        let numEdges = 0;
-        for (const v of this.vertices()) {
-            numEdges += this.outDegree(v);
-        }
-        return numEdges;
-    }
-    //-----------------------------------------------------------------
-    // MutableGraph
-    clear (): void {
-        // ComponentGraph
-        this._names.length = 0;
-        this._subpasses.length = 0;
-        // Graph Vertices
-        this._vertices.length = 0;
-    }
-    addVertex (
-        name: string,
-        subpass: RasterSubpass,
-    ): number {
-        const vert = new SubpassGraphVertex();
-        const v = this._vertices.length;
-        this._vertices.push(vert);
-        this._names.push(name);
-        this._subpasses.push(subpass);
-        return v;
-    }
-    clearVertex (v: number): void {
-        const vert = this._vertices[v];
-        // clear out edges
-        for (const oe of vert._outEdges) {
-            const target = this._vertices[oe.target as number];
-            for (let i = 0; i !== target._inEdges.length;) { // remove all edges
-                if (target._inEdges[i].target === v) {
-                    target._inEdges.splice(i, 1);
-                } else {
-                    ++i;
-                }
-            }
-        }
-        vert._outEdges.length = 0;
-
-        // clear in edges
-        for (const ie of vert._inEdges) {
-            const source = this._vertices[ie.target as number];
-            for (let i = 0; i !== source._outEdges.length;) { // remove all edges
-                if (source._outEdges[i].target === v) {
-                    source._outEdges.splice(i, 1);
-                } else {
-                    ++i;
-                }
-            }
-        }
-        vert._inEdges.length = 0;
-    }
-    removeVertex (u: number): void {
-        this._vertices.splice(u, 1);
-        this._names.splice(u, 1);
-        this._subpasses.splice(u, 1);
-
-        const sz = this._vertices.length;
-        if (u === sz) {
-            return;
-        }
-
-        for (let v = 0; v !== sz; ++v) {
-            const vert = this._vertices[v];
-            reindexEdgeList(vert._outEdges, u);
-            reindexEdgeList(vert._inEdges, u);
-        }
-    }
-    addEdge (u: number, v: number): ED | null {
-        // update in/out edge list
-        this._vertices[u]._outEdges.push(new OutE(v));
-        this._vertices[v]._inEdges.push(new OutE(u));
-        return new ED(u, v);
-    }
-    removeEdges (u: number, v: number): void {
-        const source = this._vertices[u];
-        // remove out edges of u
-        for (let i = 0; i !== source._outEdges.length;) { // remove all edges
-            if (source._outEdges[i].target === v) {
-                source._outEdges.splice(i, 1);
-            } else {
-                ++i;
-            }
-        }
-        // remove in edges of v
-        const target = this._vertices[v];
-        for (let i = 0; i !== target._inEdges.length;) { // remove all edges
-            if (target._inEdges[i].target === u) {
-                target._inEdges.splice(i, 1);
-            } else {
-                ++i;
-            }
-        }
-    }
-    removeEdge (e: ED): void {
-        const u = e.source as number;
-        const v = e.target as number;
-        const source = this._vertices[u];
-        for (let i = 0; i !== source._outEdges.length;) {
-            if (source._outEdges[i].target === v) {
-                source._outEdges.splice(i, 1);
-                break; // remove one edge
-            } else {
-                ++i;
-            }
-        }
-        const target = this._vertices[v];
-        for (let i = 0; i !== target._inEdges.length;) {
-            if (target._inEdges[i].target === u) {
-                target._inEdges.splice(i, 1);
-                break; // remove one edge
-            } else {
-                ++i;
-            }
-        }
-    }
-    //-----------------------------------------------------------------
-    // NamedGraph
-    vertexName (v: number): string {
-        return this._names[v];
-    }
-    vertexNameMap (): SubpassGraphNameMap {
-        return new SubpassGraphNameMap(this._names);
-    }
-    //-----------------------------------------------------------------
-    // PropertyGraph
-    get (tag: string): SubpassGraphNameMap | SubpassGraphSubpassMap {
-        switch (tag) {
-        // Components
-        case 'Name':
-            return new SubpassGraphNameMap(this._names);
-        case 'Subpass':
-            return new SubpassGraphSubpassMap(this._subpasses);
-        default:
-            throw Error('property map not found');
-        }
-    }
-    //-----------------------------------------------------------------
-    // ComponentGraph
-    component<T extends SubpassGraphComponent> (id: T, v: number): SubpassGraphComponentType[T] {
-        switch (id) {
-        case SubpassGraphComponent.Name:
-            return this._names[v] as SubpassGraphComponentType[T];
-        case SubpassGraphComponent.Subpass:
-            return this._subpasses[v] as SubpassGraphComponentType[T];
-        default:
-            throw Error('component not found');
-        }
-    }
-    componentMap<T extends SubpassGraphComponent> (id: T): SubpassGraphComponentPropertyMap[T] {
-        switch (id) {
-        case SubpassGraphComponent.Name:
-            return new SubpassGraphNameMap(this._names) as SubpassGraphComponentPropertyMap[T];
-        case SubpassGraphComponent.Subpass:
-            return new SubpassGraphSubpassMap(this._subpasses) as SubpassGraphComponentPropertyMap[T];
-        default:
-            throw Error('component map not found');
-        }
-    }
-    getName (v: number): string {
-        return this._names[v];
-    }
-    setName (v: number, value: string) {
-        this._names[v] = value;
-    }
-    getSubpass (v: number): RasterSubpass {
-        return this._subpasses[v];
-    }
-
-    readonly components: string[] = ['Name', 'Subpass'];
-    readonly _vertices: SubpassGraphVertex[] = [];
-    readonly _names: string[] = [];
-    readonly _subpasses: RasterSubpass[] = [];
-}
-
-export class RasterPass {
-    readonly rasterViews: Map<string, RasterView> = new Map<string, RasterView>();
-    readonly computeViews: Map<string, ComputeView[]> = new Map<string, ComputeView[]>();
-    readonly subpassGraph: SubpassGraph = new SubpassGraph();
-    width = 0;
-    height = 0;
-    readonly viewport: Viewport = new Viewport();
-    versionName = '';
-    version = 0;
-    showStatistics = false;
 }
 
 export class ComputePass {
