+++ native/cocos/renderer/pipeline/custom/NativePipelineFwd.h
@@ -45,7 +45,6 @@ class NativeComputePassBuilder;
 class NativeMovePassBuilder;
 class NativeCopyPassBuilder;
 class NativeSceneTransversal;
-struct PersistentRenderPassAndFramebuffer;
 struct RenderInstancingQueue;
 struct DrawInstance;
 struct RenderDrawQueue;
