+++ native/cocos/renderer/pipeline/custom/RenderGraphFwd.h
@@ -45,6 +45,10 @@ struct ResourceStates;
 struct ManagedBuffer;
 struct ManagedTexture;
 struct ManagedResource;
+struct RasterSubpass;
+struct SubpassGraph;
+struct RasterPass;
+struct PersistentRenderPassAndFramebuffer;
 struct ManagedTag;
 struct ManagedBufferTag;
 struct ManagedTextureTag;
@@ -54,9 +58,6 @@ struct FramebufferTag;
 struct SwapchainTag;
 struct SamplerTag;
 struct ResourceGraph;
-struct RasterSubpass;
-struct SubpassGraph;
-struct RasterPass;
 struct ComputePass;
 struct CopyPass;
 struct MovePass;
